College Project/requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.3.0
billiard==4.2.1
blinker==1.9.0
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crispy-tailwind==1.0.3
cryptography==45.0.5
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.6
django-allauth==65.11.1
django-crispy-forms==2.4
django-htmx==1.23.2
django-ratelimit==4.1.0
django-tailwind==4.2.0
ecdsa==0.19.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.116.1
Flask==3.1.1
flatbuffers==25.2.10
frozenlist==1.7.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.73.1
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
keras==3.10.0
kombu==5.5.4
libclang==18.1.1
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
multidict==6.6.4
namex==0.1.0
numpy==2.1.3
openai==1.93.0
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.0
passlib==1.7.4
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-jose==3.5.0
pytz==2025.2
redis==6.4.0
requests==2.32.5
rich==14.0.0
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.47.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tomli==2.2.1
tqdm==4.67.1
twilio==9.8.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
whitenoise==6.9.0
wrapt==1.17.2
yarl==1.20.1
